adding on to last commit, *big change for sensor initialization and reading
in sensor_read_task init we now loop through the constant at the top SENSOR_COUNT and initialize all the sensors with their respective attributes, then in the actual task loop we constant read the sensor's data and apply this data to current temp. Though there is a major issue right now, sensor 1 writes to current_temp and so does sensor 2, so the next commit will contain a method of storing two temp values and the average temp value

+++ apps/thermals/include/thermal_sensor.h
@@ -25,7 +25,7 @@ returns negative if unsuccessful (e.g. address is invalid)
 takes in a pointer to thermal_sensor_t struct, along with the address of the sensor, this will initialize
 the sensor variable correlated with the struct with the actual address of the sensor
 */
-float thermal_sensor_read(const thermal_sensor_t *sensor, float *temperature_out);
+int thermal_sensor_read(const thermal_sensor_t *sensor, float *temperature_out);
 
 
 /*
